Refactoring CyclopsExternalDamage to CyclopsDamage to include fire damage.
Added rudimentary server time syncing.
